Update pubspec.lock with all versions
