+++ lib/screens/activities/meditation_station.dart
@@ -1,0 +1,173 @@
+import 'package:flutter/material.dart';
+import 'package:audioplayers/audioplayers.dart';
+import 'dart:developer';
+
+void main() {
+  runApp(const MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const MyHomePage(title: 'CalmQuest'),
+    );
+  }
+}
+
+class AudioPlayerScreen extends StatefulWidget {
+  const AudioPlayerScreen({Key? key}) : super(key: key);
+
+  @override
+  _AudioPlayerScreenState createState() => _AudioPlayerScreenState();
+}
+
+
+class _AudioPlayerScreenState extends State<AudioPlayerScreen> {
+  late AudioPlayer _audioPlayer;
+  bool isPlaying = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _audioPlayer = AudioPlayer();
+  }
+
+  Future<void> futurePlayAudio() async {
+    String url = "https://www.soundhelix.com/examples/mp3/SoundHelix-Song-1.mp3";
+    try {
+      await _audioPlayer.play(UrlSource(url));
+      setState(() {
+        isPlaying = true;
+      });
+    } catch (e) {
+      log("Error playing audio: $e");
+    }
+  }
+
+  Future<void> _pauseAudio() async {
+    try {
+      await _audioPlayer.pause();
+      setState(() {
+        isPlaying = false;
+      });
+    } catch (e) {
+      log("Error pausing audio: $e");
+    }
+  }
+
+  Future<void> _stopAudio() async {
+    try {
+      await _audioPlayer.stop();
+      setState(() {
+        isPlaying = false;
+      });
+    } catch (e) {
+      log("Error stopping audio: $e");
+    }
+  }
+
+  @override
+  void dispose() {
+    _audioPlayer.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text("Audio Player"),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            ElevatedButton(
+              onPressed: isPlaying ? _pauseAudio : futurePlayAudio,
+              child: const Text("Pause"),
+            ),
+            ElevatedButton(
+              onPressed: _stopAudio,
+              child: const Text("Stop"),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key, required this.title});
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  String _dropdownValue = '1';
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Container(
+              width: 200,
+              height: 100,
+              decoration: BoxDecoration(
+                color: const Color(0xffEBEDFE),
+                borderRadius: BorderRadius.circular(10),
+              ),
+              child: Center(
+                child: DropdownButton<String>(
+                  value: _dropdownValue,
+                  icon: const Icon(Icons.menu),
+                  style: const TextStyle(color: Colors.black),
+                  underline: Container(
+                    height: 2,
+                    color: Colors.black,
+                  ),
+                  onChanged: (String? newValue) {
+                    setState(() {
+                      _dropdownValue = newValue!;
+                    });
+                  },
+                  items: const [
+                    DropdownMenuItem(value: '1', child: Text('Option 1')),
+                    DropdownMenuItem(value: '2', child: Text('Option 2')),
+                    DropdownMenuItem(value: '3', child: Text('Option 3')),
+                  ],
+                ),
+              ),
+            ),
+            SizedBox(height: 20),
+            ElevatedButton(
+              onPressed: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => AudioPlayerScreen()),
+                );
+              },
+              child: const Text("Go to Audio Player"),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
